src/requirements.txt
python-dotenv
Flask
requests
httpx
flask[async]